requirements.txt
fastapi
cryptography
pyjwt
uvicorn